Threat-Intelligence/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
virustotal-api==1.1.11
abuseipdb==0.1.0
OTXv2==1.5.7 